+++ EFTEMj/src/main/java/sr_eels/SimpleCoordinateCorrection.java
@@ -8,7 +8,7 @@ public class SimpleCoordinateCorrection extends CoordinateCorrector {
 
     @Override
     float calcY2n(final float x1, final float x2) throws SR_EELS_Exception {
-	final float y2n = functionWidth.getY2n(new double[] { x1, x2 });
+	final float y2n = functionWidth.getY2n(new float[] { x1, x2 });
 	return y2n;
     }
 
@@ -19,7 +19,7 @@ public class SimpleCoordinateCorrection extends CoordinateCorrector {
 
     @Override
     float calcY2(final float y1, final float y2n) {
-	final float y2 = (float) functionBorder.getY2(new double[] { y1, y2n });
+	final float y2 = functionBorder.getY2(new float[] { y1, y2n });
 	return y2;
     }
 }
